% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{compute_map_rejection_probs}
\alias{compute_map_rejection_probs}
\title{Computes the  "map" of rejection  probabilities for the Bayes  risk optimal
test of the  composite null "\eqn{\delta_x \times  \delta_y=0}" against its
alternative  "\eqn{\delta_x  \times  \delta_y\neq  0}" based  on  the  test
statistic in the real  plane.  The Bayes risk is induced  by either the 0-1
or the bounded quadratic loss function.}
\usage{
compute_map_rejection_probs(
  alpha = 0.05,
  K = 16,
  loss = c("0-1", "quadratic"),
  return_solver = FALSE
)
}
\arguments{
\item{alpha}{A positive  \code{numeric}, the wished type-I  error (default
value 0.05).}

\item{K}{An  \code{integer} parametrizing the  discretization of  the null
hypothesis  test and  of the  plane.  The  complexity of  the resulting
optimization  problem is  \eqn{O(K^2)}, so  'K' should  neither be  too
small nor too large. The default  value is 'K=16'.  We recommend 'K=64'
as a sensible value and enforce \eqn{2 \leq K \leq 128}.}

\item{loss}{A  \code{character},   either  "0-1"  (default   value)  or
"quadratic", to indicate which loss function to consider.}

\item{return_solver}{A \code{logical}, to request that the 'lpSolve' object
be returned  (if 'TRUE') or  not (if  'FALSE', default). Note  that the
'lpSolve' object can be quite big.}
}
\value{
A  '2\emph{K  x 2}K'  \code{matrix} whose  '(i,j)'  coefficient is  the
probability to  reject the null  when the  test statistic falls  in the
square \eqn{[b(i-1-K)/K,  b(i-K)/K] \times [b(j-1-K)/K,  b(j-K)/K]}. If
'return_solver'  is 'TRUE',  then the  matrix has  an attribute  called
'solver'  which is  the complete  output of  the optimization  function
'lpSolve::lp' used to determine the  probabilities.
}
\description{
Computes the  "map" of rejection  probabilities for the Bayes  risk optimal
test of the  composite null "\eqn{\delta_x \times  \delta_y=0}" against its
alternative  "\eqn{\delta_x  \times  \delta_y\neq  0}" based  on  the  test
statistic in the real  plane.  The Bayes risk is induced  by either the 0-1
or the bounded quadratic loss function.
}
\examples{
## one of the four outputs of 'compute_map_rejection_probs' stored in the package
head(map_01_0.05, c(5, 5)) 
map <- compute_map_rejection_probs(alpha = 0.05, K = 16, loss = "0-1")
plot(map)

}
