% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{map_01_0.05_t50}
\alias{map_01_0.05_t50}
\title{'Map'  of  rejection  probabilities  for the  Bayes  risk  optimal  testing
procedure, 0-1 loss function and  0.05 type-I error, without using standard
normal approximations for  the involved t-distributions with  49 degrees of
freedom}
\format{
\subsection{\code{map_01_0.05_t50}}{

A square matrix with 128 rows and columns.
}
}
\usage{
map_01_0.05_t50
}
\description{
The 'map' of rejection probabilities for the Bayes risk optimal test of the
composite null  "\eqn{\delta_x \times \delta_y=0}" against  its alternative
"\eqn{\delta_x \times \delta_y\neq  0}" based on the test  statistic in the
real plane  that is  induced by  the 0-1  loss function  for a  0.05 type-I
error. No truncation required. Does  not use standard normal approximations
for the involved t-distributions.
}
\keyword{datasets}
