% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{map_quad_0.01}
\alias{map_quad_0.01}
\title{The map of  rejection probabilities for the Bayes risk  optimal test of the
composite null  "\eqn{\delta_x \times \delta_y=0}" against  its alternative
"\eqn{\delta_x \times \delta_y\neq  0}" based on the test  statistic in the
real plane that is induced by the quadratic loss function for a 0.01 type-I
error.}
\format{
\subsection{\code{map_quad_0.01}}{

A square matrix with 128 rows and columns.
}
}
\usage{
map_quad_0.01
}
\description{
The map of  rejection probabilities for the Bayes risk  optimal test of the
composite null  "\eqn{\delta_x \times \delta_y=0}" against  its alternative
"\eqn{\delta_x \times \delta_y\neq  0}" based on the test  statistic in the
real plane that is induced by the quadratic loss function for a 0.01 type-I
error.
}
\keyword{datasets}
