% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediation.test-package.R
\docType{package}
\name{mediation.test-package}
\alias{mediation.test}
\alias{mediation.test-package}
\title{mediation.test: More Powerful Tests of the Composite Null Hypothesis Arising in Mediation Analysis}
\description{
This packages proposes more powerful tests of the composite null hypothesis arising in mediation analysis.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://achambaz.github.io/mediation.test/}
}

}
\author{
\strong{Maintainer}: Antoine Chambaz \email{antoine.chambaz@u-paris.fr}

Authors:
\itemize{
  \item Caleb Miles \email{cm3825@cumc.columbia.edu}
}

}
\keyword{internal}
