% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediation.test.R
\docType{package}
\name{mediation.test}
\alias{mediation.test}
\title{mediation.test:  More  Powerful  Tests  of the  Composite  Null  Hypothesis
Arising in Mediation Analysis}
\description{
The mediation.test package implements more  powerful tests of the composite
null hypothesis  arising in mediation  analysis. It contains  the following
functions: mediation.test, foo, bar, baz
}
