% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediation.test.R
\name{mediation_test}
\alias{mediation_test}
\title{Carries out the test of the composite null "\eqn{x \times y=0}" against its
alternative "\eqn{x  \times y\neq 0}"  based on  the test statistic  in the
real plane.}
\usage{
mediation_test(t, alpha = 0.05)
}
\arguments{
\item{t}{A  \code{vector}  consisting of  two  \code{numeric}s, the  test
statistic in the real plane.}

\item{alpha}{A positive \code{numeric}, the wished type-I error, which must
be the inverse of an integer larger  than 2 (defaults to 1/20=5\%).  If it
is not the  inverse of an integer, then it  is automatically rounded down
to the closer inverse of an integer.}
}
\value{
A list, consisting  of: \describe{ \item{t:}{a \code{vector} of two
\code{numeric}s, the test statistic;} \item{alpha:}{a \code{numeric}, the
type-I  error  (possibly  rounded  down  to  the  closer  inverse  of  an
integer);}  \item{decision:}{a \code{logical},  \code{FALSE} if  the null
hypothesis  can be  rejected for  the  alternative at  level 'alpha'  and
\code{TRUE}   otherwise;}   \item{int_pval:}{a   \code{vector}   of   two
\code{numeric}s,  a lower  bound  and  an upper  bound  on the  p-value;}
\item{pval:}{a \code{numeric}, a (random) p-value drawn uniformly between
the two aforementiond bounds.}}
}
\description{
Carries out the test of the composite null "\eqn{x \times y=0}" against its
alternative "\eqn{x  \times y\neq 0}"  based on  the test statistic  in the
real plane.
}
\examples{
(mt <- mediation_test(c(1.1, 2.2), alpha = 1/20))
plot(mt)

}
