% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediation.test.Bayes.R
\name{mediation_test_Bayes}
\alias{mediation_test_Bayes}
\title{Carries  out   the  Bayes   risk  optimal  test   of  the   composite  null
"\eqn{\delta_x    \times     \delta_y=0}"    against     its    alternative
"\eqn{\delta_x \times \delta_y\neq  0}" based on the test  statistic in the
real plane.}
\usage{
mediation_test_Bayes(t, map, truncation = 0)
}
\arguments{
\item{t}{A  \code{vector}  consisting of  two  \code{numeric}s, the  test
statistic in  the real  plane, or a  'n x 2'  \code{matrix} of  such test
statistics.}

\item{map}{The "map" of rejection  probabilities -- the 'map'  item of the
output of function \code{compute_map_rejection_probs}.}

\item{truncation}{A nonnegative  \code{numeric} used to bound the rejection
region away  from the null  hypothesis space.  Defaults to 0,  in which
case the  rejection region is minimax optimal.}
}
\value{
A list, consisting  of: \describe{ \item{t:}{a \code{vector} of two
\code{numeric}s, the test statistic, or a 'n x 2' \code{matrix} of such
test  statistics;} \item{alpha:}{a  \code{numeric}, the  type-I error;}
\item{truncation:}{a  nonnegative  \code{numeric},  used to  bound  the
rejection    region   away    from   the    null   hypothesis    space}
\item{decision:}{a  \code{vector} of  \code{logical}s, \code{FALSE}  if
the  null hypothesis  can  be  rejected for  the  alternative at  level
'alpha'  and \code{TRUE}  otherwise;}  \item{pval:}{a \code{vector}  of
\code{numeric}s,  the  p-values  of  the tests,  'NA'  in  this  case;}
\item{method:}{the  \code{character} "Bayes";}\item{map:}{The  "map" of
rejection probabilities  -- the  'map' item of  the output  of function
\code{compute_map_rejection_probs}.} }
}
\description{
Carries  out   the  Bayes   risk  optimal  test   of  the   composite  null
"\eqn{\delta_x    \times     \delta_y=0}"    against     its    alternative
"\eqn{\delta_x \times \delta_y\neq  0}" based on the test  statistic in the
real plane.
}
\details{
For details, we refer  to the technical report  "Optimal Tests of
the Composite Null Hypothesis Arising in Mediation Analysis", by
Miles & Chambaz (2021), https://arxiv.org/abs/2107.07575
}
\examples{
n <- 10
x <- MASS::mvrnorm(n, mu = c(0, 0), Sigma = diag(c(1, 1)))
delta <- matrix(stats::runif(2 * n, min = -3, max = 3), ncol = 2)
epsilon <- stats::rbinom(n, 1, 1/2)
delta <- delta * cbind(epsilon, 1 - epsilon)
x <- x + delta
(mt <- mediation_test_Bayes(x, map = map_01_0.05))
plot(mt)

}
