% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediation.test.minimax.R
\name{mediation_test_minimax}
\alias{mediation_test_minimax}
\title{Carries   out   the   minimax   optimal  test   of   the   composite   null
"\eqn{\delta_x    \times     \delta_y=0}"    against     its    alternative
"\eqn{\delta_x \times \delta_y\neq  0}" based on the test  statistic in the
real plane.}
\usage{
mediation_test_minimax(
  t,
  alpha = 0.05,
  truncation = 0,
  sample_size = Inf,
  compute_pvals = TRUE
)
}
\arguments{
\item{t}{A  \code{vector}  consisting of  two  \code{numeric}s, the  test
statistic in  the real  plane, or a  'n x 2'  \code{matrix} of  such test
statistics.}

\item{alpha}{A positive \code{numeric}, the wished type-I error.}

\item{truncation}{A nonnegative  \code{numeric} used to bound the rejection
region away  from the null  hypothesis space.  Defaults to 0,  in which
case the  rejection region is minimax optimal.}

\item{sample_size}{An \code{integer}  (larger than  one), the size  of the
sample used  to derive the  test statistic. Defaults to  'Inf', meaning
that, under the  null hypothesis, the test statistic is  drawn from the
\eqn{N_2(0,I_2)} law.  If  the integer is finite, then,  under the null
hypothesis, the test statistic is drawn from the product of two Student
laws with 'sample_size-1' degrees of freedom.}

\item{compute_pvals}{A   \code{logical}  indicating   whether  or   not
(conservative) p-values should be computed.}
}
\value{
A list, consisting  of: \describe{ \item{t:}{a \code{vector} of two
\code{numeric}s, the test statistic, or a 'n x 2' \code{matrix} of such
test  statistics;} \item{alpha:}{a  \code{numeric}, the  type-I error;}
\item{truncation:}{a  nonnegative  \code{numeric},  used to  bound  the
rejection    region   away    from   the    null   hypothesis    space}
\item{sample_size:}{an \code{integer},  the size of the  sample used to
derive   the  test   statistic}  \item{decision:}{a   \code{vector}  of
\code{logical}s, \code{FALSE}  if the  null hypothesis can  be rejected
for  the  alternative  at  level 'alpha'  and  \code{TRUE}  otherwise;}
\item{pval:}{a \code{vector}  of \code{numeric}s,  the (conservative)
p-values  of the tests;} \item{method:}{the \code{character} "minimax".}}
}
\description{
Carries   out   the   minimax   optimal  test   of   the   composite   null
"\eqn{\delta_x    \times     \delta_y=0}"    against     its    alternative
"\eqn{\delta_x \times \delta_y\neq  0}" based on the test  statistic in the
real plane.
}
\details{
For details, we refer  to the technical report  "Optimal Tests of
the Composite Null Hypothesis Arising in Mediation Analysis", by
Miles & Chambaz (2024), https://arxiv.org/abs/2107.07575
}
\examples{
n <- 10
x <- MASS::mvrnorm(2 * n, mu = c(0, 0), Sigma = diag(c(1, 1)))
delta <- matrix(stats::runif(4 * n, min = -3, max = 3), ncol = 2)
epsilon <- stats::rbinom(n, size = 1, prob = 1/2)
delta <- delta * cbind(c(epsilon, rep(1, n)),
                       c(1 - epsilon, rep(1, n)))
x <- x + delta
(mt <- mediation_test_minimax(x, alpha = 1/20))
plot(mt)

}
\seealso{
\code{\link[=mediation_test_minimax_BH]{mediation_test_minimax_BH()}}, which builds upon the present function to implement a Benjamini-Hochberg procedure to control false discovery rate.
}
