% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot.map.rejection.probabilities}
\alias{plot.map.rejection.probabilities}
\title{Plots the "map" of rejection probabilities  for the Bayes risk optimal test
of  the  composite  null  "\eqn{\delta_x \times  \delta_y=0}"  against  its
alternative  "\eqn{\delta_x  \times  \delta_y\neq  0}" based  on  the  test
statistic in the real  plane.}
\usage{
\method{plot}{map.rejection.probabilities}(x, filename = NULL, return_fig = FALSE, ...)
}
\arguments{
\item{x}{An output of \code{function} \code{}.}

\item{filename}{Either \code{NULL} (defaults) or a file  name to create on
disk.}

\item{return_fig}{A \code{logical}, to request that the  'ggplot2' object be
returned (if 'TRUE') or not (if 'FALSE').}

\item{...}{Not used.}
}
\value{
Nothing unless 'return_fig' is set to 'TRUE', in which case the function returns the 'ggplot2' object.
}
\description{
Plots the "map" of rejection probabilities  for the Bayes risk optimal test
of  the  composite  null  "\eqn{\delta_x \times  \delta_y=0}"  against  its
alternative  "\eqn{\delta_x  \times  \delta_y\neq  0}" based  on  the  test
statistic in the real  plane.
}
\examples{

map <- compute_map_rejection_probs(alpha = 0.05, K = 16, loss = "0-1")
plot(map)

}
