% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediation.test.R
\name{plot.mediation.test}
\alias{plot.mediation.test}
\title{Plots the output of \code{function} \code{mediation_test} with the rejection region.}
\usage{
\method{plot}{mediation.test}(x, filename = NULL, ...)
}
\arguments{
\item{x}{An output of \code{function} \code{mediation_test}.}

\item{filename}{Either \code{NULL} (defaults) or a file  name to create on
disk.}

\item{...}{Not used.}
}
\value{
Nothing.
}
\description{
Plots the output of \code{function} \code{mediation_test} with the rejection region.
}
\examples{
x <- MASS::mvrnorm(10, mu = c(0, 0), Sigma = diag(c(1, 1)))
delta <- matrix(stats::runif(20, min = -3, max = 3), ncol = 2)
epsilon <- stats::rbinom(10, 1, 1/2)
delta <- delta * cbind(epsilon, 1 - epsilon)
x <- x + delta
(mt <- mediation_test(x, alpha = 1/20))
plot(mt)

}
