% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediation.test.R
\name{plot3d.mediation.test}
\alias{plot3d.mediation.test}
\alias{plot3d}
\title{Plots  the   output  of  \code{function}  \code{mediation_test}   with  the
corresponding 3D power surface.}
\usage{
\method{plot3d}{mediation.test}(
  x,
  filename = NULL,
  nbins_xy = 501,
  range_delta_x = c(-6, 6),
  range_delta_y = c(-6, 6),
  ...
)
}
\arguments{
\item{x}{An output of \code{function} \code{mediation_test}.}

\item{filename}{Either \code{NULL} (defaults) or a file  name to create on
disk. If \code{NULL}, then the plot is sent to the default web browser.}

\item{nbins_xy}{An \code{integer} (chosen between 50  and 1000, defaulting
to 501), the number of bins used to discretize the along each axis.}

\item{range_delta_x, range_delta_y}{Two \code{vector}s  of \code{numeric}s
describing the  ranges of \eqn{\delta_x} and  \eqn{\delta_y}. By default,
they are both set to \code{c(-6, 6)}.}

\item{...}{Not used.}
}
\value{
Nothing.
}
\description{
Plots  the   output  of  \code{function}  \code{mediation_test}   with  the
corresponding 3D power surface.
}
\examples{
n <- 10
x <- MASS::mvrnorm(n, mu = c(0, 0), Sigma = diag(c(1, 1)))
delta <- matrix(stats::runif(2 * n, min = -3, max = 3), ncol = 2)
epsilon <- stats::rbinom(n, 1, 1/2)
delta <- delta * cbind(epsilon, 1 - epsilon)
x <- x + delta
(mt <- mediation_test(x, alpha = 1/20))
plot3d(mt)

}
