% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediation.test.R
\name{print.mediation.test}
\alias{print.mediation.test}
\title{Prints the output of \code{function} \code{mediation_test}.}
\usage{
\method{print}{mediation.test}(x, ...)
}
\arguments{
\item{x}{An output of \code{function} \code{mediation_test}.}

\item{...}{Not used.}
}
\value{
Nothing.
}
\description{
Prints the output of \code{function} \code{mediation_test}.
}
\examples{
(mt <- mediation_test(c(1.1, 2.2), alpha = 1/20))
plot(mt)

}
