% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{map_quad_0.05}
\alias{map_quad_0.05}
\title{'Map'  of  rejection  probabilities  for the  Bayes  risk  optimal  testing
procedure, quadratic  loss function and  0.05 type-I error,  using standard
normal approximations for the involved t-distributions}
\format{
\subsection{\code{map_quad_0.05}}{

A square matrix with 128 rows and columns.
}
}
\usage{
map_quad_0.05
}
\description{
The map of  rejection probabilities for the Bayes risk  optimal test of the
composite null  "\eqn{\delta_x \times \delta_y=0}" against  its alternative
"\eqn{\delta_x \times \delta_y\neq  0}" based on the test  statistic in the
real plane that is induced by the quadratic loss function for a 0.05 type-I
error. No truncation required. Uses  standard normal approximations for the
involved t-distributions.
}
\keyword{datasets}
