% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediation.test.R
\name{plot.mediation.test}
\alias{plot.mediation.test}
\title{Plots the output of \code{function} \code{mediation_test} with the rejection region.}
\usage{
\method{plot}{mediation.test}(
  x,
  filename = NULL,
  return_fig = FALSE,
  xlim = c(-4, 4),
  ylim = c(-4, 4),
  ...
)
}
\arguments{
\item{x}{An output of \code{function} \code{mediation_test}.}

\item{filename}{Either \code{NULL} (defaults) or a file  name to create on
disk.}

\item{return_fig}{A \code{logical}, to request that the  ggplot2 object be
returned (if 'TRUE') or not (if 'FALSE').}

\item{xlim, ylim}{Two \code{vectors} of \code{numeric}s,  the wished x-axis
and y-axis ranges (both default to 'c(-4, 4)').}

\item{...}{Not used.}
}
\value{
Nothing.
}
\description{
Plots the output of \code{function} \code{mediation_test} with the rejection region.
}
\examples{
n <- 10
x <- MASS::mvrnorm(n, mu = c(0, 0), Sigma = diag(c(1, 1)))
delta <- matrix(stats::runif(2 * n, min = -3, max = 3), ncol = 2)
epsilon <- stats::rbinom(n, 1, 1/2)
delta <- delta * cbind(epsilon, 1 - epsilon)
x <- x + delta
(mt <- mediation_test(x, alpha = 1/20))
plot(mt)

}
