% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediation.test.R
\name{plot.mediation.test}
\alias{plot.mediation.test}
\title{Plots the output of \code{function} \code{mediation_test}.}
\usage{
\method{plot}{mediation.test}(x, filename = NULL, ...)
}
\arguments{
\item{x}{An output of \code{function} \code{mediation_test}.}

\item{filename}{Either \code{NULL} (defaults) or a file  name to create on
disk.}

\item{...}{Not used.}
}
\value{
Nothing.
}
\description{
Plots the output of \code{function} \code{mediation_test}.
}
\examples{
(mt <- mediation_test(c(1.1, 2.2), alpha = 1/20))
plot(mt)
}
