% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.mediation.test}
\alias{print.mediation.test}
\title{Prints the output of \code{function}s \code{mediation_test_minimax} and \code{mediation_test_Bayes}.}
\usage{
\method{print}{mediation.test}(x, ...)
}
\arguments{
\item{x}{An output of \code{function} \code{mediation_test_minimax} or \code{mediation_test_Bayes}.}

\item{...}{Not used.}
}
\value{
Nothing.
}
\description{
Prints the output of \code{function}s \code{mediation_test_minimax} and \code{mediation_test_Bayes}.
}
\examples{
n <- 10
x <- MASS::mvrnorm(n, mu = c(0, 0), Sigma = diag(c(1, 1)))
delta <- matrix(stats::runif(2 * n, min = -3, max = 3), ncol = 2)
epsilon <- stats::rbinom(n, 1, 1/2)
delta <- delta * cbind(epsilon, 1 - epsilon)
x <- x + delta
(mt <- mediation_test_minimax(x, alpha = 1/20))
plot(mt)

}
